Implement LDR

+++ LC3/Instructions/LD.cs
@@ -6,7 +6,7 @@ namespace LC3.Instructions {
             var mdr = processor[Register.MemoryData];
             var register = (Register)(mdr >> 9 & 0b111);
             var value = mdr & 0b1_1111_1111;
-            processor[register] = processor.Memory[(ushort) (processor[Register.PC] + (ushort)value)];
+            processor[register] = processor.Memory[processor[Register.PC] + value];
 
             if (Program.Disassemble) {
                 Console.WriteLine($"LD\t{register}, ${value:X}");
